Move marsyfier into own class, remove references to Marsey.Serializer
