+++ Marsey/Subversion/Subverse.cs
@@ -1,18 +1,9 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
 using System.Reflection;
 using HarmonyLib;
-using Marsey.Config;
-using Marsey.Game;
 using Marsey.Game.Misc;
 using Marsey.Handbreak;
-using Marsey.Stealthsey;
 using Marsey.Misc;
 using Marsey.PatchAssembly;
-using Marsey.Serializer;
-using Marsey.Stealthsey.Reflection;
 
 namespace Marsey.Subversion;
